Adds kTLS to the gateway
