+++ Dockerfile_Gateway
@@ -1,6 +1,6 @@
 # syntax=docker/dockerfile:experimental
 
-FROM golang:1.23-alpine as dev
+FROM golang:1.25-alpine as dev
 RUN apk add --no-cache git ca-certificates upx
 RUN adduser -D appuser
 COPY . /src/
