+++ gateway/go.mod
@@ -1,24 +1,38 @@
 module gateway
 
-go 1.23.3
+go 1.25.0
 
 require (
 	github.com/cilium/ebpf v0.19.0
 	github.com/docker/docker v28.3.3+incompatible
 	github.com/gookit/slog v0.5.8
+	gitlab.com/go-extension/tls v0.0.0-20250918192917-db5d892cc1da
 )
 
 require (
+	github.com/RyuaNerin/go-krypto v1.3.0 // indirect
+	github.com/andybalholm/brotli v1.2.0 // indirect
+	github.com/blang/semver/v4 v4.0.0 // indirect
 	github.com/containerd/log v0.1.0 // indirect
+	github.com/emmansun/gmsm v0.33.0 // indirect
 	github.com/gookit/color v1.5.4 // indirect
 	github.com/gookit/goutil v0.6.18 // indirect
 	github.com/gookit/gsr v0.1.0 // indirect
+	github.com/klauspost/compress v1.18.0 // indirect
+	github.com/pedroalbanese/camellia v0.0.0-20220911183557-30cc05c20118 // indirect
+	github.com/pmorjan/kmod v1.1.1 // indirect
 	github.com/sirupsen/logrus v1.9.3 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
-	golang.org/x/sync v0.11.0 // indirect
-	golang.org/x/sys v0.31.0 // indirect
-	golang.org/x/term v0.29.0 // indirect
-	golang.org/x/text v0.22.0 // indirect
+	gitlab.com/go-extension/aes-ccm v0.0.0-20230221065045-e58665ef23c7 // indirect
+	gitlab.com/go-extension/hash v0.0.0-20250912170447-263d1d8375e4 // indirect
+	gitlab.com/go-extension/hpke v0.0.0-20250903154322-ae11394c5e06 // indirect
+	gitlab.com/go-extension/rand v0.0.0-20240303103951-707937a049b5 // indirect
+	gitlab.com/go-extension/utils v0.0.0-20250718194058-bae8b5a74647 // indirect
+	golang.org/x/crypto v0.42.0 // indirect
+	golang.org/x/sync v0.17.0 // indirect
+	golang.org/x/sys v0.36.0 // indirect
+	golang.org/x/term v0.35.0 // indirect
+	golang.org/x/text v0.29.0 // indirect
 	gotest.tools/v3 v3.5.2 // indirect
 )
