--- watcher/ca.crt
@@ -1,24 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID9zCCAt+gAwIBAgICBnUwDQYJKoZIhvcNAQELBQAwgYwxFTATBgNVBAYMDENP
-VU5UUllfQ09ERTERMA8GA1UECBMIUFJPVklOQ0UxDTALBgNVBAcTBENJVFkxEDAO
-BgNVBAkTB0FERFJFU1MxFDASBgNVBBEMC1BPU1RBTF9DT0RFMRowGAYDVQQKDBFP
-UkdBTklaQVRJT05fTkFNRTENMAsGA1UEAxMENDJDQTAeFw0yNTA4MTQxMjE1Mjda
-Fw0zNTA4MTQxMjE1MjdaMIGMMRUwEwYDVQQGDAxDT1VOVFJZX0NPREUxETAPBgNV
-BAgTCFBST1ZJTkNFMQ0wCwYDVQQHEwRDSVRZMRAwDgYDVQQJEwdBRERSRVNTMRQw
-EgYDVQQRDAtQT1NUQUxfQ09ERTEaMBgGA1UECgwRT1JHQU5JWkFUSU9OX05BTUUx
-DTALBgNVBAMTBDQyQ0EwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDa
-R6CNi9y3ovmG81y6oHGDxFPwpfYebf/lvp2nyx/WBgweZkTmwwAWwM+MLkCKe+WV
-WOugzrCSRwoT6llavDej3SpzQ0RHF25Ar8O+LaMJZvXAN18p54FWE8U8NC4faZPQ
-3fSz8sGBwfVY/z9In888pHfKjT2D63KAZu3tpCxTDngABNz/ynDOrluemWw4RjuF
-Zwog+t2UWi7/7+/RIex9QW6CFZe+p8R1nn27NuCpsz0NAJWFYftIYtYCTUoKycAy
-hFv555eVZEVZieztqWDrpM9IsF18+kk1BOgt64+8vB3wnp8Hg+Njz78Bsjh7V4Qg
-ma2ZGeIeHMRP7g18OeDRAgMBAAGjYTBfMA4GA1UdDwEB/wQEAwIChDAdBgNVHSUE
-FjAUBggrBgEFBQcDAgYIKwYBBQUHAwEwDwYDVR0TAQH/BAUwAwEB/zAdBgNVHQ4E
-FgQUsowtBSONcaH/W4/XAOhOvc7kOsUwDQYJKoZIhvcNAQELBQADggEBAJxAneeR
-LZn1kUgqsC+C/f0ibAxGZZerm56uQP5ctdeV6Cib5APxseEMY4KodERiWi6MzwT0
-ZlcmZQxaRyVR0/JXFh9q1ZxzlXwWHExUoIV8pGeAwz0A5o9U96zvQpzOtU5k28yP
-XzjdnhkmchLeBhF2+I7I7DxwL0Oztt3Hv8m5tYBGXD/P8HUryKvMKo7aI2lRvYXr
-5TJ0ujeCwv438HySkwEH2rVyr0Tzb/xszbSdl8YbH6C9Twwd0cqftVYCJlfXZoTm
-JFWjyNkCnZ8YAtjTXEBWadbXuRAsDPMOowf0Kja86IiaNClVA1ccycxMdjDOJ9D/
-XFUh1xTO6gFDJrs=
------END CERTIFICATE-----
