--- watcher/ca.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEogIBAAKCAQEA2kegjYvct6L5hvNcuqBxg8RT8KX2Hm3/5b6dp8sf1gYMHmZE
-5sMAFsDPjC5AinvllVjroM6wkkcKE+pZWrw3o90qc0NERxduQK/Dvi2jCWb1wDdf
-KeeBVhPFPDQuH2mT0N30s/LBgcH1WP8/SJ/PPKR3yo09g+tygGbt7aQsUw54AATc
-/8pwzq5bnplsOEY7hWcKIPrdlFou/+/v0SHsfUFughWXvqfEdZ59uzbgqbM9DQCV
-hWH7SGLWAk1KCsnAMoRb+eeXlWRFWYns7alg66TPSLBdfPpJNQToLeuPvLwd8J6f
-B4PjY8+/AbI4e1eEIJmtmRniHhzET+4NfDng0QIDAQABAoIBAB1IbqmOkdWw73jD
-PTVsQM5EkF5AZXw/lYzkvZG9FlBID2gkoCm1D1X+F9neepoZdPtMKpAwKk/1Z8y0
-K7iYd35sNT2xUH+ZDRlwevoQdUbVoTJgO0wL2rYJgPRKq8VbDJGAtUtOzdbyhk+i
-LoM3rT2dctShjhcchwfixxiSZb/eeYzfk+5s/hvt6yvvR/ic4XDx7AteRSIT7+NY
-A0JH+j11coZIowaIqmESqSaydjnmCQvAy3vaCdRJ6fehwwQGNptf1BkuVLdXoHux
-0lrkeVTi9lUyOKmj6j5aT+66hsXLUXUxzXCRBISbrURCZr8NDlO/tNKnFvyYTG05
-YwaIXeECgYEA4dSEnb2bbCZFxfEhvoNXEYWJ8ocPUJJbRah8PCC38C71kmye4/kz
-vgaJsAlW0+DRKnlmhBnda0USmDoCSeVIKaPWG2AKjfsPtwnb0fCvbLSjRqfF1/Ie
-WraExZCOUylcroQyxCqpdw/WMCZow2CF1iSoUeUPfyejvnnW74+OjbMCgYEA93Di
-Sn9rc0bsOK5T7NYablBc1Jf9PVUfstrVHHHkYT47jfDM2f9Y+kU9CYVHVN5sHqMn
-t5ppJK0//IOzR7knXTVHE2GnhdO9YjBmLx3a2Z7FkxsxZcARoKBggWP9VqfPgC4h
-tCA3rwSJdbN2ywRYoK7lsdJxEKDFoUdcwiYbPWsCgYBbi7IC2+UQb2djeX+LYmS+
-Vpr4NOXxbJMRZt6N5VmLRQ4JwHycv93Iz4+rS+3l0EIabMIxFOccYBDMot9go1Cg
-iKj3yfQdYEQbmCSDM4V5lLSHpOVfx+ekME8cY6A+eAr5N9/4g67U9BeIx1k6snsu
-tweAw1GH40LOqUeHL3PDXwKBgAMjNEtbiEJC8PXbCAdlw5oIfNTZRIXFkYxUCAW/
-T/UHs/udxQsymK87t4VqYjgqRdaxnOxQEes8IyaRiPFs4Jnm9grB3zzsLWeUU4df
-aon94GV4H805fT+7KN4bJXg20oO+kgDVs78NYKzUZLo4SONotFLn5zUdxMPBwuRX
-VVsXAoGAAXoXddp7wHHl5AMc0O7Yfv+/MvuNNwvGM0hBSD8vK5Wd0RvzvwXdtO7k
-rFTqpepUhmYcwp1TUWiHMW0c3z51QYFhbZyapWx+alSUFvzda7b52KgQMasmQ59c
-DFSVC4tYedf54x8vgcaygSwMwfLCXlYzu8UzGHzAFbUTu7MbJnc=
------END RSA PRIVATE KEY-----
